added flags
I added flag functionality to the program. Basically it allows you to modify input without the need to do it manually. Flags have their descriptions, long and short form.

Basic usage:

`typer -l 50 -r file1.txt file2.txt`

would randomize the imput from both files and limit words to 50

Flags have to be given before the rest of the arguments, which become tail, or they wont work.

`typer -shf file1.txt -shf file2.txt file3.txt`

would shuffle the words only in files 1 and 2 (separately) and add file3 in the end.

It is not 100% tested yet, but for now $(curl ...) doesn't work yet (need to read files or read text from termial if an argument is not a path).

Please let me know if you would even consider a functionality like this for me to work further on it.